feat(DoxyGen): doxygen config file with example comment block Fixes AREA-82
It is now possible to generate documentation using DoxyGen.

FYI, Golang and react are not meant to be used with doxygen : doxygen
is made to be used with languages like C++, C, Python and a lot of
others you can see in documentation or in the config file.

Because of this reason, I had to tell doxygen to treat golang files
like C files, which breaks the tool a little bit.

Some comments won't display until you specify `EXTRACT_ALL = YES` in
the DoxyFile. But if you do so, every single comment will appear on the
generated documentation.

I think we have to delete all useless comments so we can have a proper
doxygen generation.